@@ -11,6 +11,8 @@ struct Command {
     std::string thrust;
 
     friend std::ostream& operator<<(std::ostream& out, Command c) {
-        return out << c.target.real() << " " << c.target.imag() << " "  << c.thrust << std::endl;
+        return out << static_cast<int>(c.target.real()) << " "
+                   << static_cast<int>(c.target.imag()) << " "
+                   << c.thrust << std::endl;
     }
 };
