@@ -2,16 +2,14 @@
 
 #include "Drone.hpp"
 
-namespace {
-double scalar_product(Vector a, Vector b) {
-    return a.real()*b.real() + a.imag()*b.imag();
-}
-
-Vector map_vector(Vector a, Vector b) {
-    return scalar_product(a, b) * b / std::pow(std::abs(b), 2.0);
-}
-}
+CourseCalculator::CourseCalculator(double force) : m_correction_coefficient(force) {}
 
 Vector CourseCalculator::calculateCorrection(Drone const& drone, Position target) {
-    return - drone.speed + map_vector(drone.speed, target - drone.position);
+    auto trajectory = target - drone.position;
+    auto course = std::arg(trajectory);
+    auto angle = course - std::arg(drone.speed);
+    auto speed_parallel = std::polar(std::abs(drone.speed) * std::cos(angle), course);
+    auto speed_perpendicular = drone.speed - speed_parallel;
+
+    return -m_correction_coefficient * speed_perpendicular;
 }
